% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_sentences}
\alias{count_sentences}
\title{ellipsetxt::countSentences}
\usage{
count_sentences(txt)
}
\arguments{
\item{textblock}{: the text to analyse}
}
\value{
sentence.count : the number of sentences in the block
}
\description{
counts and returns the numbers of sentences in a text block
}
\examples{
example


}
