% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_vector_sentences}
\alias{count_vector_sentences}
\title{ellipsetxt::countVecSentences}
\usage{
count_vector_sentences(vector_of_corpus)
}
\arguments{
\item{vector_of_corpus}{the vector of text strings}
}
\value{
The number of sentences in the entire vector of corpus of text
}
\description{
counts and returns the umber of sentences in a vector of text blocks (strings)
}
\examples{
example
}
