% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_words}
\alias{count_words}
\title{ellipsetxt::wordCount}
\usage{
count_words(txt)
}
\arguments{
\item{txt}{the string}
}
\value{
an integer
}
\description{
returns the number of words in a string
}
\examples{
example



}
