% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{remove_titles}
\alias{remove_titles}
\title{ellipsetxt::cleanCorpus}
\usage{
remove_titles(txt)
}
\arguments{
\item{txt}{the vector of text strings}
}
\value{
a cleaned version of txt without M., Mr. and Dr.
}
\description{
counts and returns the umber of sentences in a vector of text blocks (strings)
}
\examples{
example
}
